requirements.txt
pyarrow
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
langchain
langchain_openai
openai
pinecone-client==2.2.4
pinecone-datasets
pypdf
python-dotenv
tiktoken
python-dotenv